requirements.txt
Django==5.2.8
channels==4.3.1
channels-redis==4.3.0
celery[redis]==5.5.3
google-genai==1.33.0
gunicorn==23.0.0
psycopg[binary]==3.2.12
python-dotenv==1.1.1
reportlab==4.4.4
uvicorn[standard]==0.38.0
whitenoise==6.11.0
